requirements.txt
torch
transformers>=4.30
datasets
evaluate
rouge-score
pyarrow
pyyaml
tqdm
accelerate
sentencepiece
pytest
pytest-mock

